fix: suppress C4267 size_t→int warning in utils.cpp Windows build
WideCharToMultiByte's cbMultiByte parameter is `int`, but
`target_length` is `size_t`. MSVC treats C4267 as an error (/WX),
so the implicit narrowing conversion aborted the build.

Cast the argument with static_cast<int> to make the conversion
explicit and silence the error.

+++ apiapp/windows/runner/utils.cpp
@@ -61,7 +61,8 @@ std::string Utf8FromUtf16(const wchar_t* utf16_string) {
   utf8_string.resize(target_length);
   int converted_length = ::WideCharToMultiByte(
       CP_UTF8, WC_ERR_INVALID_CHARS, utf16_string,
-      input_length, utf8_string.data(), target_length, nullptr, nullptr);
+      input_length, utf8_string.data(), static_cast<int>(target_length),
+      nullptr, nullptr);
   if (converted_length == 0) {
     return std::string();
   }
